+++ src/main/java/team/hanaro/hanamate/domain/MyWallet/Dto/ResponseDto.java
@@ -1,4 +1,4 @@
-package team.hanaro.hanamate.domain.MyWallet;
+package team.hanaro.hanamate.domain.MyWallet.Dto;
 
 import lombok.*;
 import team.hanaro.hanamate.entities.Account;
