+++ src/main/java/team/hanaro/hanamate/domain/MyWallet/Repository/TransactionRepository.java
@@ -1,4 +1,4 @@
-package team.hanaro.hanamate.domain.MyWallet;
+package team.hanaro.hanamate.domain.MyWallet.Repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import team.hanaro.hanamate.entities.Transactions;
@@ -7,7 +7,7 @@ import java.sql.Timestamp;
 import java.util.List;
 import java.util.Optional;
 
-public interface MyWalletTransactionsRepository extends JpaRepository<Transactions, Long> {
+public interface TransactionRepository extends JpaRepository<Transactions, Long> {
 
     Optional<List<Transactions>> findAllByWalletId(Long aLong);
