+++ src/main/java/team/hanaro/hanamate/domain/MyWallet/Repository/WalletRepository.java
@@ -1,4 +1,4 @@
-package team.hanaro.hanamate.domain.MyWallet;
+package team.hanaro.hanamate.domain.MyWallet.Repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
@@ -8,7 +8,7 @@ import team.hanaro.hanamate.entities.Wallets;
 
 import java.util.Optional;
 
-public interface MyWalletRepository extends JpaRepository<Wallets, Long> {
+public interface WalletRepository extends JpaRepository<Wallets, Long> {
 
     @Override
     Optional<Wallets> findById(Long aLong);
